+++ contracts/registry/src/lib.rs
@@ -269,9 +269,17 @@ impl Contract {
 
             key_new.issuer_id = key.issuer_id;
             key_new.class_id = key.class_id;
-            // TODO: decide if we should overwrite or panic if receipient already had a token.
-            // now we overwrite.
-            self.balances.insert(&key_new, token_id);
+            // One use can have max one toke of a (issuer, class) pair. We don't allow users
+            // to overwrite each other tokens. Recipient or sender should firstly burn his SBT
+            // to avoid conflicts.
+            if self.balances.insert(&key_new, token_id).is_some() {
+                env::panic_str(&format!(
+                    "recipient already has an SBT of issuer={}, class={}; source_token_id={}",
+                    self.issuer_by_id(key.issuer_id),
+                    key.class_id,
+                    token_id
+                ));
+            }
             self.balances.remove(key);
 
             let i_key = IssuerTokenId {
@@ -1419,11 +1427,10 @@ mod tests {
         assert_eq!(test_utils::get_logs(), log_ban);
     }
 
-    #[test]
-    fn soul_transfer_limit() {
+    fn soul_transfer_prepare() -> (VMContext, Contract) {
         let (mut ctx, mut ctr) = setup(&issuer1(), 150 * MINT_DEPOSIT);
-        let batch_metadata = mk_batch_metadata(100);
-        assert!(batch_metadata.len() == 100);
+        let batch_metadata = mk_batch_metadata(110);
+        assert_eq!(batch_metadata.len(), 110);
 
         // issuer_1
         ctr.sbt_mint(vec![(alice(), batch_metadata[..50].to_vec())]);
@@ -1431,50 +1438,51 @@ mod tests {
 
         // issuer_2
         ctx.predecessor_account_id = issuer2();
-        ctx.prepaid_gas = max_gas();
         testing_env!(ctx.clone());
-        ctr.sbt_mint(vec![(alice(), batch_metadata[50..].to_vec())]);
+        ctr.sbt_mint(vec![(alice(), batch_metadata[50..100].to_vec())]);
         assert_eq!(ctr.sbt_supply_by_owner(alice(), issuer2(), None), 50);
 
         // add more tokens to issuer_1
         ctx.predecessor_account_id = issuer1();
-        ctx.prepaid_gas = max_gas();
         testing_env!(ctx.clone());
         ctr.sbt_mint(vec![(bob(), batch_metadata[..20].to_vec())]);
         assert_eq!(ctr.sbt_supply_by_owner(bob(), issuer1(), None), 20);
 
-        ctx.prepaid_gas = max_gas();
-        testing_env!(ctx.clone());
-        ctr.sbt_mint(vec![(alice2(), batch_metadata[..20].to_vec())]);
-        assert_eq!(ctr.sbt_supply_by_owner(alice2(), issuer1(), None), 20);
+        // mint non conflicting tokens
+        ctr.sbt_mint(vec![(alice2(), batch_metadata[100..].to_vec())]);
+        assert_eq!(ctr.sbt_supply_by_owner(alice2(), issuer1(), None), 10);
 
-        ctx.prepaid_gas = max_gas();
-        testing_env!(ctx.clone());
+        testing_env!(ctx.clone()); // reset gas
         ctr.sbt_mint(vec![(carol(), batch_metadata[..20].to_vec())]);
         assert_eq!(ctr.sbt_supply_by_owner(carol(), issuer1(), None), 20);
 
-        ctx.prepaid_gas = max_gas();
         testing_env!(ctx.clone());
         ctr.sbt_mint(vec![(dan(), batch_metadata[..10].to_vec())]);
         assert_eq!(ctr.sbt_supply_by_owner(dan(), issuer1(), None), 10);
+
+        (ctx, ctr)
+    }
+
+    #[test]
+    fn soul_transfer_limit() {
+        let (mut ctx, mut ctr) = soul_transfer_prepare();
 
         // soul transfer alice->alice2
         ctx.predecessor_account_id = alice();
         ctx.prepaid_gas = max_gas();
         testing_env!(ctx.clone());
-        let limit: u32 = 25; //anything above this limit will fail due to exceeding maximum gas usage per call
+        let limit: u32 = 20; //anything above this limit will fail due to exceeding maximum gas usage per call
 
         let mut result = ctr._sbt_soul_transfer(alice2(), limit as usize);
         while !result.1 {
-            ctx.prepaid_gas = max_gas();
-            testing_env!(ctx.clone());
+            testing_env!(ctx.clone()); // reset gas
             result = ctr._sbt_soul_transfer(alice2(), limit as usize);
         }
 
         // check all the balances afterwards
         assert_eq!(ctr.sbt_supply_by_owner(alice(), issuer1(), None), 0);
         assert_eq!(ctr.sbt_supply_by_owner(alice(), issuer2(), None), 0);
-        assert_eq!(ctr.sbt_supply_by_owner(alice2(), issuer1(), None), 70);
+        assert_eq!(ctr.sbt_supply_by_owner(alice2(), issuer1(), None), 60);
         assert_eq!(ctr.sbt_supply_by_owner(alice2(), issuer2(), None), 50);
         assert_eq!(ctr.sbt_supply_by_owner(bob(), issuer1(), None), 20);
         assert_eq!(ctr.sbt_supply_by_owner(carol(), issuer1(), None), 20);
@@ -1483,56 +1491,35 @@ mod tests {
 
     #[test]
     #[should_panic(expected = "HostError(GasLimitExceeded)")]
-    fn soul_transfer_exceeded_limit() {
-        let (mut ctx, mut ctr) = setup(&issuer1(), 150 * MINT_DEPOSIT);
-        let batch_metadata = mk_batch_metadata(100);
-        assert!(batch_metadata.len() == 100);
-
-        // issuer_1
-        ctr.sbt_mint(vec![(alice(), batch_metadata[..50].to_vec())]);
-        assert_eq!(ctr.sbt_supply_by_owner(alice(), issuer1(), None), 50);
-
-        // issuer_2
-        ctx.predecessor_account_id = issuer2();
-        ctx.prepaid_gas = max_gas();
-        testing_env!(ctx.clone());
-        ctr.sbt_mint(vec![(alice(), batch_metadata[50..].to_vec())]);
-        assert_eq!(ctr.sbt_supply_by_owner(alice(), issuer2(), None), 50);
-
-        // add more tokens to issuer_1
-        ctx.predecessor_account_id = issuer1();
-        ctx.prepaid_gas = max_gas();
-        testing_env!(ctx.clone());
-        ctr.sbt_mint(vec![(bob(), batch_metadata[..20].to_vec())]);
-        assert_eq!(ctr.sbt_supply_by_owner(bob(), issuer1(), None), 20);
-
-        ctx.prepaid_gas = max_gas();
-        testing_env!(ctx.clone());
-        ctr.sbt_mint(vec![(alice2(), batch_metadata[..20].to_vec())]);
-        assert_eq!(ctr.sbt_supply_by_owner(alice2(), issuer1(), None), 20);
-
-        ctx.prepaid_gas = max_gas();
-        testing_env!(ctx.clone());
-        ctr.sbt_mint(vec![(carol(), batch_metadata[..20].to_vec())]);
-        assert_eq!(ctr.sbt_supply_by_owner(carol(), issuer1(), None), 20);
-
-        ctx.prepaid_gas = max_gas();
-        testing_env!(ctx.clone());
-        ctr.sbt_mint(vec![(dan(), batch_metadata[..10].to_vec())]);
-        assert_eq!(ctr.sbt_supply_by_owner(dan(), issuer1(), None), 10);
+    fn soul_transfer_exceeded_gas_limit() {
+        let (mut ctx, mut ctr) = soul_transfer_prepare();
 
         // soul transfer alice->alice2
         ctx.predecessor_account_id = alice();
         ctx.prepaid_gas = max_gas();
         testing_env!(ctx.clone());
-        let limit: u32 = 30; //anything above this limit will fail due to exceeding maximum gas usage per call
+        let limit: u32 = 30;
+        ctr._sbt_soul_transfer(alice2(), limit as usize);
+    }
 
-        let mut result = ctr._sbt_soul_transfer(alice2(), limit as usize);
-        while !result.1 {
-            ctx.prepaid_gas = max_gas();
-            testing_env!(ctx.clone());
-            result = ctr._sbt_soul_transfer(alice2(), limit as usize);
-        }
+    #[test]
+    #[should_panic(
+        expected = "recipient already has an SBT of issuer=sbt.n, class=1; source_token_id=1"
+    )]
+    fn soul_transfer_conflict() {
+        let (mut ctx, mut ctr) = soul_transfer_prepare();
+
+        // alice1 already has SBT of (issuer1, 1), so let's mint it to alice2 and try soul_transfer
+        let batch_metadata = mk_batch_metadata(1);
+        ctx.predecessor_account_id = issuer1();
+        testing_env!(ctx.clone());
+        ctr.sbt_mint(vec![(alice2(), batch_metadata)]);
+
+        ctx.predecessor_account_id = alice();
+        ctx.prepaid_gas = max_gas();
+        testing_env!(ctx.clone());
+        let limit: u32 = 30;
+        ctr._sbt_soul_transfer(alice2(), limit as usize);
     }
 
     #[test]